refactor-components * Move CloseButton to components directory
