+++ src/components/CloseButton/CloseButton.scss
