+++ src/components/CloseButton/CloseButton.tsx
@@ -2,11 +2,11 @@ import React from 'react'
 import { CloseIcon } from '@entur/icons'
 import './CloseButton.scss'
 
-interface Props {
+interface CloseButtonProps {
     onClick: () => void
 }
 
-const CloseButton = ({ onClick }: Props): JSX.Element => (
+const CloseButton: React.FC<CloseButtonProps> = ({ onClick }) => (
     <button className="close-button" onClick={onClick}>
         <CloseIcon className="icon" />
         Lukk
