+++ src/components/Modals/LoginModal/EmailSent/EmailSent.tsx
@@ -4,7 +4,7 @@ import { GridContainer, GridItem } from '@entur/grid'
 import { PrimaryButton } from '@entur/button'
 import { Heading3, Paragraph } from '@entur/typography'
 import { ModalType } from '../LoginModal'
-import { CloseButton } from '../CloseButton/CloseButton'
+import { CloseButton } from '../../../CloseButton/CloseButton'
 import Check from '../../../../assets/images/check.png'
 import retinaCheck from '../../../../assets/images/check@2x.png'
