+++ src/components/Modals/LoginModal/LoginModal.tsx
@@ -2,6 +2,7 @@ import React, { useState, useEffect } from 'react'
 import type { User } from 'firebase/auth'
 import { useToast } from '@entur/alert'
 import { Modal } from '@entur/modal'
+import { CloseButton } from '../../CloseButton/CloseButton'
 import { useUser } from '../../../UserProvider'
 import { usePrevious } from '../../../utils'
 import { EmailLogin } from './EmailLogin/EmailLogin'
@@ -9,7 +10,6 @@ import { LoginOptions } from './LoginOptions/LoginOptions'
 import { Signup } from './Signup/Signup'
 import { ResetPassword } from './ResetPassword/ResetPassword'
 import { EmailSent } from './EmailSent/EmailSent'
-import { CloseButton } from './CloseButton/CloseButton'
 import './LoginModal.scss'
 
 export type LoginCase =
