+++ src/components/Modals/MineTavlerModal/MineTavlerModal.tsx
@@ -7,7 +7,7 @@ import { PrimaryButton, SecondaryButton } from '@entur/button'
 import { useUser } from '../../../UserProvider'
 import { useSettings } from '../../../settings/SettingsProvider'
 import { LoginModal } from '../LoginModal/LoginModal'
-import { CloseButton } from '../LoginModal/CloseButton/CloseButton'
+import { CloseButton } from '../../CloseButton/CloseButton'
 import sikkerhetBom from '../../../assets/images/sikkerhet_bom.png'
 import retinaSikkerhetBom from '../../../assets/images/sikkerhet_bom@2x.png'
 import { getDocumentId } from '../../../utils'
